+++ Changelog.md
@@ -3,6 +3,34 @@
 Changelog [format](http://keepachangelog.com/en/1.0.0/)
 
 
+## Margit Bridge 3.23.1
+
+### Fixed
+* BRIDGE-488: Restore previous RemoveFromMailbox behavior (< 3.23.0). Drafts moved to trash should not be perma-deleted.
+
+
+## Margit Bridge 3.23.0
+
+### Added
+* BRIDGE-432: MacOS26 new icon customization support.
+* BRIDGE-312: New option to display the current Bridge version in CLI mode.
+* BRIDGE-453: Display UI spinner when Bridge has no active internet connection to the server.
+
+### Changed
+* BRIDGE-477: Log additional data for Human Verification requests.
+* BRIDGE-255: Updated macOS tray icon setting descriptions.
+* BRIDGE-442: Include the CC header when importing messages.
+* BRIDGE-451: Do not call message unlabel endpoints when moving to/from folders.
+
+### Fixed
+* BRIDGE-454: Prevent crashes during synchronization metadata stage (out-of-bounds).
+* BRIDGE-474: Prevent panic on Windows 10 1809, and resolve runtime crash on Windows 10 22H2 related to Hardware Key authentication.
+* BRIDGE-444: Upgrade account banner dismisses correctly when signing in.
+* BRIDGE-431: Resolve Ubuntu autostart issue, multiple escape characters included during overwrite.
+* BRIDGE-337: Display relevent account when selecting from tray menu.
+* BRIDGE-452: Resolve second-password prompt cancellation leading to wrong Login screen.
+
+
 ## Laviolette Bridge 3.22.0
 
 ### Added